refactor(theme): simplify i18n tokens
