+++ packages/theme/src/components/Layout.astro
@@ -67,7 +67,7 @@ const containerClass = wide ? "max-w-3xl lg:max-w-6xl" : "max-w-3xl"
 ---
 
 <!DOCTYPE html>
-<html lang={t.htmlLocale}>
+<html lang={t.lang}>
 <head>
   <meta charset="UTF-8" />
   <meta name="viewport" content="width=device-width, initial-scale=1.0" />
