+++ packages/theme/src/i18n/locales/en.ts
@@ -1,10 +1,10 @@
 import { format } from "date-fns"
 
-import type { Tokens } from "../types"
+import type { Tokens } from "../tokens"
 
 const tokens: Tokens = {
-  htmlLocale: "en",
-  segmenterLocale: "en",
+  lang: "en",
+  formatDate: (date) => format(date, "PP"),
 
   search: "Search",
   searchPlaceholder: "Search across notes and posts",
@@ -19,6 +19,16 @@ const tokens: Tokens = {
   searchNoResults: (q) => `No notes found for "${q}".`,
 
   about: "About",
+  explore: "Explore",
+  archive: "Archive",
+  back: "Back",
+  next: "Next",
+  previous: "Previous",
+  morePostsBackToAll: "Back to all posts",
+  backToAllNotes: "Back to all notes",
+  backToYear: (year) => `Back to ${year}`,
+  backToMonth: (date) => `Back to ${format(date, "MMMM yyyy")}`,
+  backToAllSeries: "Back to all series",
 
   notes: "Notes",
   posts: "Posts",
@@ -31,62 +41,46 @@ const tokens: Tokens = {
   seriesCountSubtitle: (n) => `${n} total`,
   postsInSeriesUnit: (n) => `${n} posts in this series`,
 
-  pinnedPosts: "Pinned posts",
-  recentPosts: "Recent posts",
-  explore: "Explore",
-  archive: "Archive",
-  yearsLabel: "Years",
-  yearLabel: (year) => `${year}`,
-  monthYearLabel: (date) => format(date, "MMMM yyyy"),
-  monthShortLabel: (date) => format(date, "MMM"),
-  otherSide: "The other side of me",
-
   noteDraft: "Draft",
-  postDraft: "Draft",
   noteUntitled: "Untitled note",
   notesEmpty: "No notes found.",
-  postsEmpty: "No posts found.",
-  seriesEmpty: "No series found.",
-  readPost: "Read the full post",
-  partOfSeries: "Part of the series:",
-  oldestFirst: "Oldest first",
-
-  notFoundBack: "Back to home",
-  back: "Back",
-  imageTitle: (caption) => (caption ? `Image: ${caption}` : "Image"),
-
-  builtWithPrefix: "Built with",
-  builtWithSuffix: "",
-  pageLabel: "Page",
-
-  next: "Next",
-  previous: "Previous",
   noteIdLabel: (n) => `Note #${n}`,
-  morePostsBackToAll: "Back to all posts",
-  backToAllNotes: "Back to all notes",
-  backToYear: (year) => `Back to ${year}`,
-  backToMonth: (date) => `Back to ${format(date, "MMMM yyyy")}`,
-  backToAllSeries: "Back to all series",
   moreNotes: (n) => `More notes (${n} total)`,
-
   notesIndexTitle: "Notes",
   notesYearTitle: (year) => `Notes in ${year}`,
   notesMonthTitle: (date) => `Notes in ${format(date, "MMMM yyyy")}`,
   notesDayTitle: (date) => `Notes on ${format(date, "PP")}`,
+
+  pinnedPosts: "Pinned posts",
+  recentPosts: "Recent posts",
+  postDraft: "Draft",
+  postsEmpty: "No posts found.",
+  readPost: "Read the full post",
   postsTitle: "Posts",
+
+  seriesEmpty: "No series found.",
   seriesTitle: "Series",
-  tableOfContents: "On this page",
+  partOfSeries: "Part of the series:",
+  oldestFirst: "Oldest first",
+
+  yearsLabel: "Years",
+  yearLabel: (year) => `${year}`,
+  monthYearLabel: (date) => format(date, "MMMM yyyy"),
+  monthShortLabel: (date) => format(date, "MMM"),
 
   heatmapTitle: "Writing journey",
   heatmapEmpty: (date) => `No writing on ${format(date, "MMM d, yyyy")}`,
   heatmapNoteOnly: (n, date) => `${n} ${n === 1 ? "note" : "notes"} on ${format(date, "MMM d, yyyy")}`,
   heatmapPostOnly: (n, date) => `${n} ${n === 1 ? "post" : "posts"} on ${format(date, "MMM d, yyyy")}`,
-  heatmapNoteAndPost: (notes, posts, date) =>
-    `${notes} ${notes === 1 ? "note" : "notes"} with ${posts} ${posts === 1 ? "post" : "posts"} on ${format(date, "MMM d, yyyy")}`,
+  heatmapNoteAndPost: (notes, posts, date) => `${notes} ${notes === 1 ? "note" : "notes"} with ${posts} ${posts === 1 ? "post" : "posts"} on ${format(date, "MMM d, yyyy")}`,
 
-  formatDate: (date) => format(date, "PP"),
-  formatDateLong: (date) => format(date, "PP"),
-  formatDateForGroup: (date) => format(date, "PP"),
+  tableOfContents: "On this page",
+  otherSide: "The other side of me",
+  notFoundBack: "Back to home",
+  imageTitle: (caption) => (caption ? `Image: ${caption}` : "Image"),
+  builtWithPrefix: "Built with",
+  builtWithSuffix: "",
+  pageLabel: "Page",
 }
 
 export default tokens
