+++ packages/theme/src/i18n/locales/zh.ts
@@ -1,10 +1,10 @@
 import { format } from "date-fns"
 
-import type { Tokens } from "../types"
+import type { Tokens } from "../tokens"
 
 const tokens: Tokens = {
-  htmlLocale: "zh",
-  segmenterLocale: "zh",
+  lang: "zh",
+  formatDate: (date) => format(date, "yyyy 年 M 月 d 日"),
 
   search: "搜索",
   searchPlaceholder: "搜索手记与文章",
@@ -19,6 +19,16 @@ const tokens: Tokens = {
   searchNoResults: (q) => `没有找到与「${q}」相关的手记。`,
 
   about: "关于",
+  explore: "探索",
+  archive: "归档",
+  back: "返回",
+  next: "下一则",
+  previous: "上一则",
+  morePostsBackToAll: "返回全部文章",
+  backToAllNotes: "返回全部手记",
+  backToYear: (year) => `返回 ${year} 年`,
+  backToMonth: (date) => `返回 ${format(date, "M 月")}`,
+  backToAllSeries: "返回全部系列",
 
   notes: "手记",
   posts: "文章",
@@ -31,62 +41,46 @@ const tokens: Tokens = {
   seriesCountSubtitle: (n) => `共 ${n} 个`,
   postsInSeriesUnit: (n) => `系列文章共 ${n} 篇`,
 
-  pinnedPosts: "置顶文章",
-  recentPosts: "近期文章",
-  explore: "探索",
-  archive: "归档",
-  yearsLabel: "年份",
-  yearLabel: (year) => `${year} 年`,
-  monthYearLabel: (date) => format(date, "yyyy 年 M 月"),
-  monthShortLabel: (date) => format(date, "M 月"),
-  otherSide: "我的另一面",
-
   noteDraft: "草稿",
-  postDraft: "草稿",
   noteUntitled: "无题手记",
   notesEmpty: "暂无手记。",
-  postsEmpty: "暂无文章。",
-  seriesEmpty: "暂无系列。",
-  readPost: "阅读全文",
-  partOfSeries: "连载于系列：",
-  oldestFirst: "按时间顺序排列",
-
-  notFoundBack: "返回首页",
-  back: "返回",
-  imageTitle: (caption) => (caption ? `图片：${caption}` : "图片"),
-
-  builtWithPrefix: "使用",
-  builtWithSuffix: "构建",
-  pageLabel: "页面",
-
-  next: "下一则",
-  previous: "上一则",
   noteIdLabel: (n) => `第 ${n} 则手记`,
-  morePostsBackToAll: "返回全部文章",
-  backToAllNotes: "返回全部手记",
-  backToYear: (year) => `返回 ${year} 年`,
-  backToMonth: (date) => `返回 ${format(date, "M 月")}`,
-  backToAllSeries: "返回全部系列",
   moreNotes: (n) => `更多手记 (共 ${n} 则)`,
-
   notesIndexTitle: "手记",
   notesYearTitle: (year) => `${year} 年手记`,
   notesMonthTitle: (date) => `${format(date, "yyyy 年 M 月")}手记`,
   notesDayTitle: (date) => format(date, "yyyy 年 M 月 d 日手记"),
+
+  pinnedPosts: "置顶文章",
+  recentPosts: "近期文章",
+  postDraft: "草稿",
+  postsEmpty: "暂无文章。",
+  readPost: "阅读全文",
   postsTitle: "文章",
+
+  seriesEmpty: "暂无系列。",
   seriesTitle: "系列",
-  tableOfContents: "目录",
+  partOfSeries: "连载于系列：",
+  oldestFirst: "按时间顺序排列",
+
+  yearsLabel: "年份",
+  yearLabel: (year) => `${year} 年`,
+  monthYearLabel: (date) => format(date, "yyyy 年 M 月"),
+  monthShortLabel: (date) => format(date, "M 月"),
 
   heatmapTitle: "写作历程",
   heatmapEmpty: (date) => `${format(date, "yyyy 年 M 月 d 日")}：无写作`,
   heatmapNoteOnly: (n, date) => `${format(date, "yyyy 年 M 月 d 日")}：${n} 则手记`,
   heatmapPostOnly: (n, date) => `${format(date, "yyyy 年 M 月 d 日")}：${n} 篇文章`,
-  heatmapNoteAndPost: (notes, posts, date) =>
-    `${format(date, "yyyy 年 M 月 d 日")}：${notes} 则手记 和 ${posts} 篇文章`,
+  heatmapNoteAndPost: (notes, posts, date) => `${format(date, "yyyy 年 M 月 d 日")}：${notes} 则手记 和 ${posts} 篇文章`,
 
-  formatDate: (date) => format(date, "yyyy 年 M 月 d 日"),
-  formatDateLong: (date) => format(date, "yyyy 年 M 月 d 日"),
-  formatDateForGroup: (date) => format(date, "yyyy 年 M 月 d 日"),
+  tableOfContents: "目录",
+  otherSide: "我的另一面",
+  notFoundBack: "返回首页",
+  imageTitle: (caption) => (caption ? `图片：${caption}` : "图片"),
+  builtWithPrefix: "使用",
+  builtWithSuffix: "构建",
+  pageLabel: "页面",
 }
 
 export default tokens
